refactor: update imports and restructure route setup in app and admin modules
